bug: reverted /subscribed page
I had to revert the /subscribed page to be the thing that adds the stripe customer Id to the profile, the stripe customer id is not available until the customer finishes the transaction. Now it should work up until the point that the user subscribes, if the user is actually a subscriber needs more testing.